refactor: revisão das rotas de feedback concluídas

+++ Backend/src/use-cases/feedbacks/delete-feedback-use-case.ts
@@ -1,6 +1,6 @@
-import { Feedback } from "@prisma/client"
-import { FeedbacksRepository } from "src/repositories/feedbacks-repository"
-import { ResourceNotFoundError } from "../errors/resource-not-found-error"
+import { Feedback } from "@prisma/client";
+import { FeedbacksRepository } from "src/repositories/feedbacks-repository";
+import { ResourceNotFoundError } from "../errors/resource-not-found-error";
 
 interface DeleteFeedbackUseCaseRequest {
     feedbackId: string;
@@ -14,12 +14,14 @@ export class DeleteFeedbackUseCase {
     constructor(private readonly feedbackRepository: FeedbacksRepository) {}
 
     async execute({ feedbackId }: DeleteFeedbackUseCaseRequest): Promise<DeleteFeedbackUseCaseResponse> {
-        const feedback = await this.feedbackRepository.delete(feedbackId)
+        const existingFeedback = await this.feedbackRepository.findById(feedbackId);
 
-        if (!feedback) {
-            throw new ResourceNotFoundError()
+        if (existingFeedback === null) {
+            throw new ResourceNotFoundError();
         }
 
-        return { feedback }
+        const deletedFeedback = await this.feedbackRepository.delete(feedbackId);
+
+        return { feedback: deletedFeedback } as DeleteFeedbackUseCaseResponse;
     }
 }
